first draft of PoE class
